SBERDOMA-63 refactor: move Ticket/schema to @domains
