+++ apps/condo/domains/Ticket/schema/Ticket.js
@@ -1,15 +1,13 @@
 const { Text, Relationship, Integer, DateTimeUtc, Checkbox } = require('@keystonejs/fields')
-const { JSON_UNKNOWN_VERSION_ERROR } = require('../../constants/errors')
 
 const access = require('@core/keystone/access')
-const { ORGANIZATION_OWNED_FIELD } = require('../_common')
 const { GQLListSchema } = require('@core/keystone/schema')
 const { Json, AutoIncrementInteger } = require('@core/keystone/fields')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 
-const { SENDER_FIELD, DV_FIELD } = require('../_common')
-const { hasRequestAndDbFields } = require('../../utils/validation.utils')
-const { JSON_EXPECT_OBJECT_ERROR, DV_UNKNOWN_VERSION_ERROR, STATUS_UPDATED_AT_ERROR } = require('../../constants/errors')
+const { SENDER_FIELD, DV_FIELD, ORGANIZATION_OWNED_FIELD } = require('../../../schema/_common')
+const { hasRequestAndDbFields } = require('../../../utils/validation.utils')
+const { JSON_EXPECT_OBJECT_ERROR, DV_UNKNOWN_VERSION_ERROR, STATUS_UPDATED_AT_ERROR, JSON_UNKNOWN_VERSION_ERROR } = require('../../../constants/errors')
 
 const ACCESS_TO_ALL = {
     read: true,
