+++ apps/condo/domains/Ticket/schema/Ticket.test.js
@@ -1,9 +1,9 @@
 const faker = require('faker')
-const { makeClientWithProperty } = require('../Property/Property.test')
+const { makeClientWithProperty } = require('../../../schema/Property/Property.test')
 const { NUMBER_RE, UUID_RE, DATETIME_RE } = require('@core/keystone/test.utils')
 const OPEN_STATUS_ID ='6ef3abc4-022f-481b-90fb-8430345ebfc2'
 
-const { Ticket } = require('../../gql/Ticket')
+const { Ticket } = require('../../../gql/Ticket')
 
 const UNKNOWN_CLASSIFIER_ID = '4f4b43d5-0951-425c-9428-945dc6193361'
 const OTHER_SOURCE_ID = '7da1e3be-06ba-4c9e-bba6-f97f278ac6e4'
