+++ apps/condo/domains/Ticket/schema/TicketStatus.js
@@ -2,8 +2,8 @@ const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 const { Text, Select } = require('@keystonejs/fields')
 
-const { SENDER_FIELD, DV_FIELD } = require('../_common')
-const { COMMON_AND_ORGANIZATION_OWNED_FIELD } = require('../_common')
+const { SENDER_FIELD, DV_FIELD } = require('../../../schema/_common')
+const { COMMON_AND_ORGANIZATION_OWNED_FIELD } = require('../../../schema/_common')
 
 const READ_ONLY_ACCESS = {
     read: true,
