+++ apps/condo/domains/Ticket/schema/index.js
